+++ index.html
@@ -66,7 +66,7 @@
         </form>
       </section>
     </header>
-    <main>
+    <main id="main">
       <section id="tags">
         <div class="filter-container">
           <button
