+++ src/style.css
@@ -92,12 +92,17 @@ body {
   stroke: black;
 }
 
+#main {
+  margin: 0 100px;
+}
+
 #tags {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 20px 0 0 100px;
+  margin-top: 50px;
   gap: 60px;
+  width: 100%;
 }
 
 .filters-btn {
@@ -138,7 +143,6 @@ body {
 }
 
 #nbr-recipes {
-  margin-right: 100px;
   font-family: "Anton", sans-serif;
   font-size: 21px;
   font-style: normal;
@@ -149,13 +153,18 @@ body {
 
 #recipes-section {
   display: flex;
-  flex-direction: column;
-  margin: 50px 100px 200px;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  margin: 50px 0 200px;
+  gap: 50px;
 }
 
 .card-recipes {
   width: 380px;
   height: 731px;
+  flex-shrink: 0;
 }
 
 .link-recipes {
@@ -170,6 +179,7 @@ body {
   color: black;
   text-decoration: none;
   object-fit: contain;
+  border-radius: 21px;
 }
 
 .img-recipes {
@@ -199,8 +209,9 @@ body {
   margin: 0 25px;
 }
 
-.recipes-div {
-  margin: 0 25px;
+.recipes-div,
+.ingredients-div {
+  padding: 0 25px;
   display: flex;
   flex-direction: column;
   gap: 15px;
@@ -217,6 +228,42 @@ body {
 
 .recipes-desc {
   color: #1b1b1b;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  height: 76px;
+}
+
+.ingredients-box {
+  display: flex;
+  width: 100%;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  gap: 21px;
+}
+
+.ingredient-div {
+  flex-basis: 45%;
+}
+
+.ingredient-name {
+  color: var(--Black, #1b1b1b);
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 500;
+}
+
+.ingredient-quantity-unit {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.ingredient-quantity,
+.ingredient-unit {
+  color: var(--Grey, #7a7a7a);
   font-size: 14px;
   font-style: normal;
   font-weight: 400;
